Moved the Solicitor Case creation into Helper file and added change state for solicitor journey
